\name{rggm-package}
\alias{rggm-package}
\alias{rggm}
\docType{package}
\title{
  Fitting Gaussian concentrate matrix models with a robustified method
}
\description{
  This package provides some robustified methods that are originally
  implemented in \code{ggm} package.
}
\details{
\tabular{rl}{
Package: \tab rggm\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2006-11-17\cr
License: \tab GPL 2 and newer\cr
}
  You can fit a Gaussian concentration matrix model with \code{rob.fitConGraph}.
}
\author{
Masashi Miyamura

Maintainer: Masashi Miyamura <miyamura@sigmath.es.osaka-u.ac.jp>
}
\references{
  Masashi Miyamura and Yutaka Kano, 2006, Robust Gaussian graphical
  modeling, Journal of Multivariate Analysis, 97, 7, 1525-1550
}
\keyword{package}
\seealso{
\code{\link[ggm:ggm-package]{ggm}}
}
