% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_ggm.R
\name{rmgaussian}
\alias{rmgaussian}
\title{Generate multivariate Gaussian observations}
\usage{
rmgaussian(n, means, covariance)
}
\arguments{
\item{n}{the sample size}

\item{means}{a vector of means}

\item{covariance}{a variacen-covariance matrix}
}
\value{
a matrix with n rows
}
\description{
This function generates multivariate Gaussian ove
}
