\name{rob.delta}
\alias{rob.delta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{An internal function}
\description{
  No description
}
\usage{
rob.delta(amat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{amat}{An adjacent matrix}
}
\author{Masashi Miyamura}
\seealso{\code{\link{rob.asymv.sigma}}}
\keyword{multivariate}% __ONLY ONE__ keyword per line
\keyword{utilities}% __ONLY ONE__ keyword per line
