% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_sbm.R
\name{rWeightSBM}
\alias{rWeightSBM}
\title{Sample random weights on edges of a binary Stochastic Block Model}
\usage{
rWeightSBM(anSBM, family = c("gaussian", "poisson"), theta)
}
\arguments{
\item{anSBM}{SBM sampled from the rSBM function with nbBlock}

\item{family}{character describing the distribution used for the weigths}

\item{theta}{list embedding parameters required for the distribution of the weights. See details.

Elements in the \code{theta} should be named as follows, depending on the argument \code{family}, with
dimension matching the number of blocks in the original SBM:
\itemize{
  \item{Gaussian}{\code{theta$mu}, a (nbBlock x nbBlock) matrix of means; \code{theta$sigma}, a (nbBlock x nbBlock) matrix of standa deviation}
  \item{Poisson}{\code{theta$lambda}, , a (nbBlock x nbBlock) matrix of means.}
  \item{Others}{zoon available.}
}}
}
\value{
an SBM weight weigthed edges
}
\description{
This function draws weights according to a user-defined distribution for the edges of an existing binary SBM.
}
\examples{
## graph parameters
nbNodes  <- 90
blockProp <- c(.5, .25, .25)   # group proportions
nbBlock   <- length(blockProp) # number of blocks
connectParam <- diag(.4, nbBlock) + 0.05 # connectivity matrix: affiliation network

## Graph Sampling
mySBM <- rSBM(nbNodes, connectParam, blockProp)

## Sampling Gaussian weights
mu_within   <- 4 ; sigma_within  <- .5
mu_between  <- 2 ; sigma_between <- .5
theta <- list()
theta$mu    <- matrix(mu_between   , nbBlock, nbBlock); diag(theta$mu)    <- mu_within    # means
theta$sigma <- matrix(sigma_between, nbBlock, nbBlock); diag(theta$sigma) <- sigma_within # sd

mySBM <- rWeightSBM(mySBM, "gaussian", theta)

hist(igraph::E(mySBM)$weight,  breaks = sqrt(igraph::gsize(mySBM)))

}
