% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_sbm.R
\name{rWeightSBM}
\alias{rWeightSBM}
\title{Sample random weights on edges of a binary Stochastic Block Model}
\usage{
rWeightSBM(anSBM, family = c("gaussian", "poisson", "laplace"), theta)
}
\arguments{
\item{anSBM}{SBM sampled from the rSBM function with nbBlock}

\item{family}{character describing the distribution used for the weigths}

\item{theta}{list embedding parameters required for the distribution of the weights. Either "gaussian", "poisson" or "laplace". See details.

Elements in the \code{theta} should be named as follows, depending on the argument \code{family}, with
dimension matching the number of blocks in the original SBM:
\itemize{
  \item{Gaussian}{\code{theta$mu}, a (nbBlock x nbBlock) matrix of means; \code{theta$sigma}, a (nbBlock x nbBlock) matrix of standard deviations}
  \item{Laplace}{\code{theta$m}, a (nbBlock x nbBlock) matrix of location parameters; \code{theta$s}, a (nbBlock x nbBlock) matrix of scale parameters}
  \item{Poisson}{\code{theta$lambda}, a (nbBlock x nbBlock) matrix of means.}
  \item{Others}{soon available.}
}}
}
\value{
an SBM weight weigthed edges
}
\description{
This function draws weights according to a user-defined distribution for the edges of an existing binary SBM.
}
\examples{
## graph parameters
nbNodes  <- 90
blockProp <- c(.5, .25, .25)   # group proportions
nbBlock   <- length(blockProp) # number of blocks
connectParam <- diag(.4, nbBlock) + 0.05 # connectivity matrix: affiliation network

## Graph Sampling
mySBM <- rSBM(nbNodes, connectParam, blockProp)

## Sampling Gaussian weights
mu_within   <- 4 ; sigma_within  <- .5
mu_between  <- 2 ; sigma_between <- .5
theta <- list()
theta$mu    <- matrix(mu_between   , nbBlock, nbBlock); diag(theta$mu)    <- mu_within    # means
theta$sigma <- matrix(sigma_between, nbBlock, nbBlock); diag(theta$sigma) <- sigma_within # sd

mySBM_Gaussian <- rWeightSBM(mySBM, "gaussian", theta)
hist(igraph::E(mySBM_Gaussian)$weight,  breaks = sqrt(igraph::gsize(mySBM_Gaussian)))

## Sampling Laplace weights
m_within   <- 4 ; s_within  <- .5
m_between  <- 2 ; s_between <- .5
theta <- list()
theta$m <- matrix(m_between, nbBlock, nbBlock); diag(theta$m) <- m_within # location parameter
theta$s <- matrix(s_between, nbBlock, nbBlock); diag(theta$s) <- s_within # scale parameters

mySBM_Laplace <- rWeightSBM(mySBM, "laplace", theta)
hist(igraph::E(mySBM_Laplace)$weight,  breaks = sqrt(igraph::gsize(mySBM_Laplace)))

## Sampling Poisson weights
lambda_within   <- 4
lambda_between  <- 4
theta <- list()
# mean/variance parameter
theta$lambda <- matrix(lambda_between, nbBlock, nbBlock)
diag(theta$lambda) <- lambda_within

mySBM_Poisson <- rWeightSBM(mySBM, "poisson", theta)
hist(igraph::E(mySBM_Poisson)$weight,  breaks = sqrt(igraph::gsize(mySBM_Poisson)))

}
